complete

+++ Continuous delivery for Intune/Install-CDforIntune.ps1
@@ -110,14 +110,43 @@ function Install-SC {
 
 function Install-HKLM {
     Param(
-        `$URL
+        `$URL,
+        `$detection
     )
-    `$TempHKLMFile = `$env:TEMP + "\TempHKLM.reg"
-    Remove-Item `$TempHKLMFile -Force | Out-Null
-    Invoke-WebRequest -Uri `$URL -OutFile `$TempHKLMFile
-    `$Arguments = "/s `$TempHKLMFile"
-    Start-Process "regedit.exe" -ArgumentList `$Arguments -Wait
-    Remove-Item `$TempHKLMFile -Force
+    `$runDetectionRule = Invoke-Expression -Command `$detection
+    If (!(`$runDetectionRule -eq `$true)) {
+        `$TempHKLMFile = `$env:TEMP + "\TempHKLM.reg"
+        Remove-Item `$TempHKLMFile -Force | Out-Null
+        Invoke-WebRequest -Uri `$URL -OutFile `$TempHKLMFile
+        `$Arguments = "/s `$TempHKLMFile"
+        Start-Process "regedit.exe" -ArgumentList `$Arguments -Wait
+        Remove-Item `$TempHKLMFile -Force
+    }
+}
+
+function Install-HKCU {
+    Param(
+        `$URL,
+        `$detection
+    )
+    `$runDetectionRule = Invoke-Expression -Command `$detection
+    
+    If (!(`$runDetectionRule -eq `$true)) {
+        `$TempHKCUFile = `$env:TEMP + "\TempHKCU.reg"
+        Remove-Item `$TempHKCUFile -Force | Out-Null
+        Invoke-WebRequest -Uri `$URL -OutFile `$TempHKCUFile
+        `$regfile = Get-Content `$TempHKCUFile
+        `$hives = Get-ChildItem -Path REGISTRY::HKEY_USERS | Select-Object -ExpandProperty Name
+        foreach (`$hive in `$hives) {
+            if (!(`$hive -like "*_Classes")) {
+                `$newregfile = `$regfile -replace "HKEY_CURRENT_USER",`$hive
+                Set-Content -Path `$TempHKCUFile -Value `$newregfile
+                `$Arguments = "/s `$TempHKCUFile"
+                Start-Process "regedit.exe" -ArgumentList `$Arguments -Wait
+            }
+        }
+        Remove-Item `$TempHKCUFile -Force
+    }
 }
 
 function Install-AdvancedApplication {
@@ -189,10 +218,21 @@ Invoke-WebRequest -Uri "https://raw.githubusercontent.com/Forsbakk/Intune-Applic
 `$HKLMFiles = Get-Content `$HKLMFileConf | ConvertFrom-Json
 
 foreach (`$hklmfile in `$HKLMFiles) {
-    Install-HKLM -URL `$hklmfile.URL
+    Install-HKLM -URL `$hklmfile.URL -detection `$hklmfile.detection
 }
 
 Remove-Item `$HKLMFileConf -Force
+
+
+`$HKCUFileConf = `$env:TEMP + "\HKCUFileConfig.JSON"
+Invoke-WebRequest -Uri "https://raw.githubusercontent.com/Forsbakk/Intune-Application-Installers/master/Continuous%20delivery%20for%20Intune/Registry/config.json" -OutFile `$RegFileConf
+`$HKCUFiles = Get-Content `$HKCUFileConf | ConvertFrom-Json
+
+foreach (`$hkcufile in `$HKCUFiles) {
+    Install-HKCU -URL `$hkcufile.URL -detection `$hkcufile.detection
+}
+
+Remove-Item `$HKCUFileConf -Force
 
 
 `$SCConfig = `$env:TEMP + "\SCConfig.JSON"
